buenas noches

+++ src/Season.java
@@ -55,14 +55,14 @@ public class Season {
             Console.printCreatingNewSeason();
             EditSeasonFile.createFile(this.teamFileName);
             EditSeasonFile.createFile(this.fixtureFileName);
-            EditSeasonFile.saveFile(this.teamFileName);
-            EditSeasonFile.saveFile(this.fixtureFileName);
+            EditSeasonFile.saveSeasonFile(this.teamsList, this.teamFileName);
+            EditFixtureFile.saveFixturesFile(this.fixturesList, this.fixtureFileName);
         } else {
             Console.printLoadingSeason(this.name);
             this.teamsList = EditSeasonFile.readTeamsFirstTime(this.teamFileName);
             this.numberOfTeams = this.teamsList.size();
             this.fixturesList = EditFixtureFile.readFixturesFirstTime(this.fixtureFileName);
-            System.out.println("the number of teams is: " + this.teamsList.size());
+            Console.printNumberOfTeams(this.teamsList.size());
         }
 
     }
@@ -70,6 +70,7 @@ public class Season {
     public int wellcome() {
         Console.printWellcomeMessage(this.name);
         int selected = this.scanner.nextInt();
+        System.out.println("\n");
         return selected;
     }
 
@@ -79,46 +80,66 @@ public class Season {
         Team team = new Team(name);
         this.numberOfTeams++;
         this.teamsList.add(team);
-        EditSeasonFile.saveTeams(this.teamsList, this.teamFileName);
+        EditFixtureFile.saveFixturesFile(this.fixturesList, this.fixtureFileName);
+        EditSeasonFile.saveSeasonFile(this.teamsList, this.teamFileName);
     }
 
     public void generateFixture() {
-        
-        this.fixturesList.clear();
-        int n = this.teamsList.size() - 1;
+        Console.printWariningMessage();
+        char option = this.scanner.next().charAt(0);
 
-        for (int i = 1; i < (n + 1); i++) {
-            int counter = 0;
-            ArrayList<Game> games = new ArrayList<>();
-            Fixture fixture = new Fixture(i);
-            if (i % 2 == 0) {
-                counter ++;
-                games.add(new Game(this.teamsList.get(i - 1), this.teamsList.get(n), fixture.id, counter));
-            } else {
-                counter ++;
-                games.add(new Game(this.teamsList.get(n), this.teamsList.get(i - 1), fixture.id, counter));
+        if (option == 'y' || option == 'Y') {
+            this.fixturesList.clear();
+            for (Team team : this.teamsList) {
+                team.resetTeam();
             }
+            if (this.teamsList.size() != 0) {
+                this.fixturesList.clear();
+                int n = this.teamsList.size() - 1;
 
-            for (int k = 1; k < (n + 1) / 2; k++) {
-                int tmp = (i + k) % n;
-                int teamA = tmp == 0 ? n : tmp;
-                tmp = ((i - k % n) + n) % n;
-                int teamB = tmp == 0 ? n : tmp;
-                if (k % 2 != 0) {
-                    counter ++;
-                    games.add(new Game(this.teamsList.get(teamA - 1), this.teamsList.get(teamB - 1), fixture.id, counter));
-                } else {
-                    counter ++;
-                    games.add(new Game(this.teamsList.get(teamB - 1), this.teamsList.get(teamA - 1), fixture.id, counter));
+                for (int i = 1; i < (n + 1); i++) {
+                    int counter = 0;
+                    ArrayList<Game> games = new ArrayList<>();
+                    Fixture fixture = new Fixture(i);
+                    if (i % 2 == 0) {
+                        counter++;
+                        games.add(new Game(this.teamsList.get(i - 1), this.teamsList.get(n), fixture.id, counter));
+                    } else {
+                        counter++;
+                        games.add(new Game(this.teamsList.get(n), this.teamsList.get(i - 1), fixture.id, counter));
+                    }
+
+                    for (int k = 1; k < (n + 1) / 2; k++) {
+                        int tmp = (i + k) % n;
+                        int teamA = tmp == 0 ? n : tmp;
+                        tmp = ((i - k % n) + n) % n;
+                        int teamB = tmp == 0 ? n : tmp;
+                        if (k % 2 != 0) {
+                            counter++;
+                            games.add(new Game(this.teamsList.get(teamA - 1), this.teamsList.get(teamB - 1), fixture.id,
+                                    counter));
+                        } else {
+                            counter++;
+                            games.add(new Game(this.teamsList.get(teamB - 1), this.teamsList.get(teamA - 1), fixture.id,
+                                    counter));
+                        }
+                    }
+                    fixture.games = games;
+                    this.fixturesList.add(fixture);
                 }
+
+                for (Team team : this.teamsList) {
+                    team.resetTeam();
+                }
+                EditFixtureFile.deleteFile(this.fixtureFileName);
+                EditFixtureFile.saveFixturesFile(this.fixturesList, this.fixtureFileName);
+                Console.printFixturesOnTerminal(this.fixturesList);
+            } else {
+                Console.printRegisterTeams();
             }
-            fixture.games = games;
-            this.fixturesList.add(fixture);
         }
-        EditFixtureFile.deleteFile(this.fixtureFileName);
-        EditFixtureFile.saveFixtures(this.fixturesList, this.fixtureFileName);
-        Console.printFixturesOnTerminal(this.fixturesList);   
-        
+
+        return;
 
     }
 
@@ -131,7 +152,7 @@ public class Season {
             int gameId = this.scanner.nextInt();
             Game game = findGame(fixtureId, gameId);
             if (game.played == true) {
-                System.out.println("This game has been already played. Do you want to update it? Y/N:");
+                Console.printGameAlreadyPlayed();
                 char option = this.scanner.next().charAt(0);
 
                 if (option == 'y' || option == 'Y') {
@@ -152,18 +173,24 @@ public class Season {
             game = this.updateGame(goalsFirstTeam, goalsSecondTeam, true, gameId, fixtureId);
             this.fixturesList.get(fixtureId - 1).games.set(gameId - 1, game);
             this.updateTeams(firstTeam, secondTeam, goalsFirstTeam, goalsSecondTeam);
-            EditFixtureFile.saveFixtures(this.fixturesList, this.fixtureFileName);
-            EditSeasonFile.saveTeams(this.teamsList, this.teamFileName);
-        }else {
-            System.out.println("Please generate fixtures first");
+            EditFixtureFile.saveFixturesFile(this.fixturesList, this.fixtureFileName);
+            EditSeasonFile.saveSeasonFile(this.teamsList, this.teamFileName);
+        } else {
+            Console.printGenerateFixtures();
         }
-        
+
     }
 
     public void showTable() {
-        this.teamsList.sort(Comparator.comparing(Team::getPoints).thenComparing(Team::getScoredGoals).thenComparing(Team::getRecievedGoals).reversed());
-        Console.printScoreboard(this.name);
-        Console.printTeamsOnTerminal(this.teamsList);
+        if (this.teamsList.size() != 0) {
+            this.teamsList.sort(Comparator.comparing(Team::getPoints).thenComparing(Team::getScoredGoals)
+                    .thenComparing(Team::getRecievedGoals).reversed());
+            Console.printScoreboard(this.name);
+            Console.printTeamsOnTerminal(this.teamsList);
+        } else {
+            Console.printRegisterTeams();
+        }
+
     }
 
     public Game findGame(int fixtureId, int gameId) {
@@ -171,25 +198,25 @@ public class Season {
     }
 
     public Game updateGame(int goalsTeam1, int goalsTeam2, boolean played, int gameId, int fixtureId) {
-        Game game = this.fixturesList.get(fixtureId - 1).games.get(gameId -1);
+        Game game = this.fixturesList.get(fixtureId - 1).games.get(gameId - 1);
         game.goalsTeam1 = goalsTeam1;
         game.goalsTeam2 = goalsTeam2;
         game.played = played;
         return game;
     }
 
-    public void updateTeams (Team team1, Team team2, int goalsTeam1, int goalsTeam2) {
+    public void updateTeams(Team team1, Team team2, int goalsTeam1, int goalsTeam2) {
 
         if (goalsTeam1 > goalsTeam2) {
-            
-            team1.setPoints(team1.getPoints() + 3);
-        } 
 
-        if(goalsTeam2 > goalsTeam1) {
+            team1.setPoints(team1.getPoints() + 3);
+        }
+
+        if (goalsTeam2 > goalsTeam1) {
             team2.setPoints(team2.getPoints() + 3);
         }
 
-        if(goalsTeam1 == goalsTeam2) {
+        if (goalsTeam1 == goalsTeam2) {
             team1.setPoints(team1.getPoints() + 1);
             team1.setPoints(team2.getPoints() + 1);
         }
@@ -202,15 +229,14 @@ public class Season {
         team2.setScoredGoals(team2.getScoredGoals() + goalsTeam2);
         team2.setrecievedGoals(team2.getRecievedGoals() + goalsTeam1);
 
-        Optional<Team> match = this.teamsList.stream().filter(team -> team.getName().equals(team1.getName())).
-        findFirst();
+        Optional<Team> match = this.teamsList.stream().filter(team -> team.getName().equals(team1.getName()))
+                .findFirst();
 
-        Optional<Team> match2 = this.teamsList.stream().filter(team -> team.getName().equals(team2.getName())).
-        findFirst();
+        Optional<Team> match2 = this.teamsList.stream().filter(team -> team.getName().equals(team2.getName()))
+                .findFirst();
 
         Team firstTeam = match.get();
         Team secondTeam = match2.get();
-        System.out.println("indice del betis es:");
         this.teamsList.set(this.teamsList.indexOf(firstTeam), team1);
         this.teamsList.set(this.teamsList.indexOf(secondTeam), team2);
     }
@@ -220,17 +246,17 @@ public class Season {
         Team team2 = game.team2;
 
         if (game.goalsTeam1 > game.goalsTeam2) {
-            
-            team1.setPoints(- 3);
-        } 
 
-        if(game.goalsTeam2 > game.goalsTeam1) {
-            team2.setPoints(- 3);
+            team1.setPoints(-3);
         }
 
-        if(game.goalsTeam1 == game.goalsTeam2) {
-            team1.setPoints(- 1);
-            team1.setPoints(- 1);
+        if (game.goalsTeam2 > game.goalsTeam1) {
+            team2.setPoints(-3);
+        }
+
+        if (game.goalsTeam1 == game.goalsTeam2) {
+            team1.setPoints(-1);
+            team1.setPoints(-1);
         }
 
         team1.unsetGamesPlayed();
@@ -241,19 +267,16 @@ public class Season {
         team2.setScoredGoals(-game.goalsTeam2);
         team2.setrecievedGoals(-game.goalsTeam1);
 
-        Optional<Team> match = this.teamsList.stream().filter(team -> team.getName().equals(team1.getName())).
-        findFirst();
+        Optional<Team> match = this.teamsList.stream().filter(team -> team.getName().equals(team1.getName()))
+                .findFirst();
 
-        Optional<Team> match2 = this.teamsList.stream().filter(team -> team.getName().equals(team2.getName())).
-        findFirst();
+        Optional<Team> match2 = this.teamsList.stream().filter(team -> team.getName().equals(team2.getName()))
+                .findFirst();
 
         Team firstTeam = match.get();
         Team secondTeam = match2.get();
         this.teamsList.set(this.teamsList.indexOf(firstTeam), team1);
         this.teamsList.set(this.teamsList.indexOf(secondTeam), team2);
-        this.teamsList.sort(Comparator.comparing(Team::getPoints).reversed());
-
-        
     }
 
 }
